Email Spam Prediction using ML,Python
An interesting data science task which is useful for all email users. The primary goal was to create a system that can automatically classify email messages as either "spam" or "ham" (not spam). Trained the model using MultinomialNaiveBayes classification and SVM, to show the difference between both the methods.Conclusion from it is that MultinomialNB gave more better result than SVM. Hence using that classifier, took few new examples and tested them.